+++ proteinworkshop/datasets/utils.py
@@ -7,6 +7,8 @@ import os
 import os.path
 import pathlib
 import tarfile
+import logging
+import time
 from typing import List, Optional
 
 import biotite.database.rcsb as rcsb
@@ -38,6 +40,8 @@ def flatten_dir(dir: os.PathLike):
                 print(f"Could not move {os.path.join(dirpath, filename)}")
 
 
+
+
 def download_pdb_mmtf(
     mmtf_dir: pathlib.Path,
     ids: Optional[List[str]] = None,
@@ -67,10 +71,26 @@ def download_pdb_mmtf(
         all_id_query = rcsb.FieldQuery("struct.title")
         pdb_ids = rcsb.search(all_id_query)
         pdb_ids = [pdb_id.lower() for pdb_id in pdb_ids]
+    else:
+        pdb_ids = ids
 
     # Name for download directory
     if not os.path.isdir(mmtf_dir):
         os.mkdir(mmtf_dir)
+
+    def catched_rcsb_fetch(pdb_id, max_retries=5, retry_interval=100):
+        data = None
+        for retry_idx in range(max_retries):
+            try:
+                data = rcsb.fetch(pdb_id, format="mmtf", target_path=mmtf_dir)
+                break
+            except Exception as e:
+                logging.warn(f"Attempt {retry_idx} to fetch pdb_id {pdb_id} failed: {e}")
+            time.sleep(retry_interval/1000)
+            
+        if data is None:
+            logging.warn(f"Failed to fetch pdb_id {pdb_id}")
+        return data
 
     # Download all PDB IDs with parallelized HTTP requests
     with concurrent.futures.ThreadPoolExecutor(max_workers=10) as executor:
@@ -82,7 +102,7 @@ def download_pdb_mmtf(
                 f"Submitting PDB download request for {pdb_id}"
             )
             futures.append(
-                executor.submit(rcsb.fetch, pdb_id, "mmtf", mmtf_dir)
+                executor.submit(catched_rcsb_fetch, pdb_id)
             )
         pbar = tqdm(concurrent.futures.as_completed(futures))
         for request_index, future in enumerate(pbar):