+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 
 * Adds `InverseSquareRoot` LR scheduler [#71](https://github.com/a-r-j/ProteinWorkshop/pull/71)
 
+### Command
+* Adds `--force-cuda-version` to `workshop install` [#78](https://github.com/a-r-j/ProteinWorkshop/pull/78)
+
 
 ### 0.2.5 (28/12/2023)